requirements.txt
streamlit
azure-storage-blob>=12.14.1
azure-storage-queue>=12.7.0
azure-ai-documentanalysis>=5.3.0
azure-core>=1.30.0
requests
reportlab
python-dotenv
numpy
pandas
scikit-learn
joblib
azure-functions
